CM-1250: Rename elasticmanager to scheduler
