+++ src/scheduler/elasticsearch/transformers/park/operatingDates.js
